Add CGBase.h and move CoreGraphics datatypes to it, fixing up headers in the process. Fix #458.
