+++ Frameworks/CoreFoundation/Base.subproj/CFBase.h
@@ -680,147 +680,6 @@ CFTypeRef CFMakeCollectable(CFTypeRef cf) CF_AUTOMATED_REFCOUNT_UNAVAILABLE;
 
 // WINOBJC: Start. Include old WinObjC additions to CFBase.h. Some might not really belong here but are still needed by other dependencies.
 #include <math.h>
-#ifdef __LP64__
-typedef double CGFloat;
-#define CGFLOAT_MIN DBL_MIN
-#define CGFLOAT_MAX DBL_MAX
-#else
-typedef float CGFloat;
-#define CGFLOAT_MIN FLT_MIN
-#define CGFLOAT_MAX FLT_MAX
-#endif
-
-#ifdef __cplusplus
-struct CGPoint {
-    CGFloat x, y;
-
-    CGPoint operator+(const CGPoint& v) const {
-        CGPoint ret = { x + v.x, y + v.y };
-        return ret;
-    }
-
-    CGPoint operator-() const {
-        CGPoint ret = { -x, -y };
-        return ret;
-    }
-
-    CGPoint operator-(const CGPoint& v) const {
-        CGPoint ret = { x - v.x, y - v.y };
-        return ret;
-    }
-
-    CGPoint operator-=(const CGPoint& v) {
-        x -= v.x;
-        y -= v.y;
-        return *this;
-    }
-
-    CGPoint operator+=(const CGPoint& v) {
-        x += v.x;
-        y += v.y;
-        return *this;
-    }
-
-    bool operator==(const CGPoint& v) {
-        return x == v.x && y == v.y;
-    }
-
-    bool operator!=(const CGPoint& v) {
-        return x != v.x || y != v.y;
-    }
-
-    CGPoint operator*(float scalar) const {
-        CGPoint ret = { x * scalar, y * scalar };
-        return ret;
-    }
-
-    CGPoint operator*=(float scalar) {
-        x *= scalar;
-        y *= scalar;
-        return *this;
-    }
-
-    CGPoint operator/(float scalar) const {
-        CGPoint ret = { x / scalar, y / scalar };
-        return ret;
-    }
-
-    CGPoint operator/=(float scalar) {
-        x /= scalar;
-        y /= scalar;
-        return *this;
-    }
-
-    CGFloat distSq(const CGPoint& v) const {
-        CGPoint delta = *this - v;
-        return delta.dot(delta);
-    }
-
-    CGFloat dist(const CGPoint& v) const {
-        return sqrt(distSq(v));
-    }
-
-    bool distGr(const CGPoint& v, CGFloat dist) const {
-        return distSq(v) > dist * dist;
-    }
-
-    CGFloat lenSq() const {
-        return dot(*this);
-    }
-
-    CGFloat len() const {
-        return sqrt(lenSq());
-    }
-
-    bool lenGe(float len) {
-        return lenSq() >= len * len;
-    }
-
-    bool lenLe(float len) {
-        return lenSq() < len * len;
-    }
-
-    CGFloat dot(const CGPoint& v) const {
-        return v.x * x + v.y * y;
-    }
-
-    CGPoint normalized() const {
-        float l = len();
-
-        if (l != 0) {
-            CGPoint ret = { x / l, y / l };
-            return ret;
-        } else {
-            CGPoint ret = { 0, 0 };
-            return ret;
-        }
-    }
-
-    static inline CGPoint point(float x, float y) {
-        CGPoint ret;
-
-        ret.x = x;
-        ret.y = y;
-
-        return ret;
-    }
-};
-#else
-typedef struct CGPoint {
-    CGFloat x;
-    CGFloat y;
-} CGPoint;
-#endif
-
-typedef struct CGSize {
-    CGFloat width;
-    CGFloat height;
-} CGSize;
-
-typedef struct CGRect {
-    CGPoint origin;
-    CGSize size;
-} CGRect;
 
 #include <CoreFoundation/CoreFoundationExport.h>
 #include <MacTypes.h>
