+++ include/CoreGraphics/CGGeometry.h
@@ -1,6 +1,6 @@
 /* Copyright (c) 2006-2007 Christopher J. W. Lloyd
 
-Copyright (c) 2016 Microsoft Corporation. All rights reserved.
+Copyright (c) Microsoft. All rights reserved.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the
 "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish,
@@ -17,8 +17,7 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. */
 #pragma once
 
 #import <CoreGraphics/CoreGraphicsExport.h>
-#import <Foundation/Foundation.h>
-#import <CoreFoundation/CFBase.h>
+#import <CoreGraphics/CGBase.h>
 #import <CoreFoundation/CFDictionary.h>
 #import <stdbool.h>
