+++ include/CoreGraphics/CGPDFObject.h
@@ -1,6 +1,6 @@
 //******************************************************************************
 //
-// Copyright (c) 2016 Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft. All rights reserved.
 //
 // This code is licensed under the MIT License (MIT).
 //
@@ -16,7 +16,7 @@
 #pragma once
 
 #import <CoreGraphics/CoreGraphicsExport.h>
-#import <CoreFoundation/CFBase.h>
+#import <CoreGraphics/CGBase.h>
 
 typedef enum {
     kCGPDFObjectTypeNull = 1,
