+++ include/CoreImage/CIFilter.h
@@ -1,6 +1,6 @@
 //******************************************************************************
 //
-// Copyright (c) 2016 Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft. All rights reserved.
 //
 // This code is licensed under the MIT License (MIT).
 //
@@ -17,6 +17,7 @@
 
 #import <CoreImage/CoreImageExport.h>
 #import <Foundation/NSObject.h>
+#import <CoreGraphics/CGBase.h>
 
 @class NSString;
 @class NSDictionary;
